Requestor statistics component added and connected to statistics tab
